Corebian/function-sql/requirements.txt
google-cloud-discoveryengine==0.9.1
google-cloud-aiplatform==1.47.0
google-cloud-dialogflow-cx==1.21.0
google-cloud-storage==2.9.0
google-cloud-bigquery==3.20.1
fuzzysearch==0.7.3
functions-framework==3.*
beautifulsoup4==4.12.2
pypdf==3.11.1
pandas==2.0.0
numpy==1.24.3
db-dtypes==1.2.0